requirements.txt
numpy
faiss-cpu
fastapi
sentence_transformers
transformers
pydantic
python-multipart
datasets
uvicorn
torch
tqdm
matplotlib